Changed to make the residuals plot separate and easier to see. Added in creation of all four clovers.
fitAll.bash:
Changed the line numbers that get seded. When creating a new folder to
start the analysis, we take the source directory to be the number of
clovers to create the data files for.

plot.gp:
This guy has been modified to plot the residuals in their own plot beneath
the main plot for the fit. This removes this issue with the scales being
mismatched and not being able to see the residuals for larger plot ranges.

Signed-off-by: S.V. Paulauskas <stanpaulauskas@gmail.com>