Creates output to copy directly to cal.txt Made things self-contained. Fixed error with label printing. Moved readme to plain text.
README.md -> README
Updated the readme to have the latest ifnormation on how to run things.

fitAll.bash -> energyCal.bash
Updated so that it no longer requires the external function call to the
makePdf script.  It will also set the position of the fitted equation to
the maximum gamma energy in the source file. Finally, it will also now
create a series of files formatted to be copied directly into a cal.txt.

plot.gp
Changes reflect those mentioned before for the label printing. Also, fixed
some margins so that things do not have so much negative space around the
plots.

Signed-off-by: S.V. Paulauskas <stanpaulauskas@gmail.com>